fix(#637): route promo-code endpoint into the right module + workspace tinting
Two regressions from the previous attempt, found via per-issue Cloud Run
inspection:

1. /api/teachers/me/promo-code returned 404 — the route was added to
   backend/routers/teachers.py but Python imports the package directory
   backend/routers/teachers/ (which shadows the file). The endpoint was
   never registered with FastAPI even though the deploy built the latest
   commit. Move it into the proper sub-module (teachers/profile_ops.py)
   alongside /me, /me/roles, /me/password, and drop the dead copy from
   teachers.py.

2. Sidebar didn't tint per workspace. Add a 5-color palette
   (Personal=indigo + 4 org colors keyed by org id hash) memoized off
   mode + selectedOrganization, and apply the active workspace's colors
   to the bottom workspace+token block (background, label, progress bar).
   Include workspaceColor and tokenInfo in the SidebarContent useMemo deps
   so the sidebar actually re-renders on workspace switch.

After redeploy demo@duotopia.com's profile will hit the auto-create path
on first load — no manual DB seed needed.
